requirements.txt
flask==3.0.0
python-docx==1.1.0
openai==1.12.0
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
